requirements.txt
pandas>=2.0.0
plotly>=5.18.0
kaleido>=0.2.1
jinja2>=3.1.0
weasyprint>=60.0
openai>=1.0.0
python-dotenv>=1.0.0
rich>=13.7.0
flask>=3.0.0
gunicorn>=21.0.0
resend>=1.0.0
